requirements.txt
pandas
python-dotenv
scikit-learn
beautifulsoup4
requests
playwright
streamlit
langdetect

# LLM provider SDKs
openai
anthropic
google-generativeai
